Misc cleanup and minor refactor
