+++ Lua/LuabindObjectWrapper.cpp
@@ -4,7 +4,7 @@
 #include "luabind/object.hpp"
 
 namespace RTE {
-	
+
 /////////////////////////////////////////////////////////////////////////////////////////////////////////////////////////////
 
 	LuabindObjectWrapper::~LuabindObjectWrapper() {
